feat(kernel,tools,llm): scheduling, host introspection, escalation channels, LLM compat
- Schedule actions: persistent recurring/once jobs with policy guard, control tool, channel sink
- Host introspection: host-package-install, network-sockets, open-files, system-mounts/services drivers
- Escalation: channel fan-out sink, expanded approval workflow surface
- LLM compat: anthropic/ollama/custom adapter improvements, tool helper hardening, retry tuning
- Kernel: agent inbox + messages, sub-agent commands, intent validator, kernel_action surface, run_loop
- Tools: agent-manual pagination overhaul, describe-tool/list-tools/search-tools rework, schedule-once revamp
- Web: chat in-flight tracking, chat handler streaming + templates, task detail + notification UI updates
- CLI: provider command expansion, agent command additions
- Types: schedule, context, tool, task additions to support new flows

+++ crates/agentos-hal/src/drivers/network_sockets.rs
@@ -253,6 +253,6 @@ mod tests {
             .iter()
             .find(|s| s.local_addr.contains(&port.to_string()))
             .unwrap();
-        assert_eq!(our_socket.pid, Some(std::process::id() as u32));
+        assert_eq!(our_socket.pid, Some(std::process::id()));
     }
 }
